pyproject.toml
[project]
name = "bookleaf-2"
version = "0.1.0"
description = "Automated AI book cover validation system for BookLeaf Publishing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.190.0",
    "google-auth-oauthlib>=1.2.4",
    "google-generativeai>=0.8.6",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pyairtable>=3.3.0",
    "pydantic>=2.12.5",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "pdf2image",
]
